refactoring: 3/n - making base components global
