+++ components/global/BaseFootnote.vue
