+++ components/global/BaseLogo.vue
