+++ components/global/BaseThanks.vue
