--- plugins/globals.js
@@ -1,10 +1,0 @@
-import Vue from 'vue'
-import BaseMetaListTools from '@/components/Base/BaseMetaListTools'
-import BaseMetaList from '@/components/Base/BaseMetaList'
-import BaseCallout from '@/components/Base/BaseCallout'
-import BaseThanks from '@/components/Base/BaseThanks'
-
-Vue.component('base-meta-list-tools', BaseMetaListTools)
-Vue.component('base-meta-list', BaseMetaList)
-Vue.component('base-callout', BaseCallout)
-Vue.component('base-thanks', BaseThanks)
